Merged PR 680: Fixed deployment errors
Pipeline now runs as expected, which is as follows:
1. Creates app service in Azure
2. Builds .NET app
3. Deploys .NET app to app service

At this point, I believe that you should be good to go, but please let me know if further improvements/modifications are needed. @<Baseem Missaghi> and I will continue to watch for any Azure compliance issues and will make changes to the ARM template as necessary to resolve those, but otherwise we will not make any changes to this repo.